Updated method for dealing with bad label encoding feats
Bad feats are features with values in the test set that are not in the
training set.
Old method for handling this was fitting the label encoder with a union of
the training and test set.
New method for handling this is to remove bad features and only encode
good features.